python/requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_ecs_patterns